requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
exceptiongroup==1.2.0
greenlet==3.0.3
iniconfig==2.0.0
packaging==23.2
playwright==1.42.0
pluggy==1.4.0
pyee==11.0.1
pytest==8.1.1
python-dotenv==1.0.1
tomli==2.0.1
typing_extensions==4.10.0
flake8==7.0.0
charset-normalizer==3.3.2
pycodestyle==2.11.1
pyflakes==3.2.0